[Decomp][Spell] Cohesion split: extract SpellEffectObjectCombat handlers to SpellEffectObjectCombat.cpp
